+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.3] - Unreleased
+## [1.0.3] - 2021-06-09
 
 ### Deprecated
 - Deprecated `NklKst\TheSportsDb\Entity\Table\Entry`, use `NklKst\TheSportsDb\Entity\Table\Standing` instead